# backend/docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/tasks/{id}:
    get:
      tags:
      - task-controller
      operationId: getTaskById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
    put:
      tags:
      - task-controller
      operationId: updateTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
    delete:
      tags:
      - task-controller
      operationId: softDeleteTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/habits/{id}:
    put:
      tags:
      - habit-controller
      operationId: updateHabit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HabitResponse"
    delete:
      tags:
      - habit-controller
      operationId: deleteHabit
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HabitResponse"
  /api/habit-completions/{id}:
    put:
      tags:
      - habit-completion-controller
      operationId: updateCompletion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitCompletionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HabitCompletionResponse"
    delete:
      tags:
      - habit-completion-controller
      operationId: deleteCompletion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HabitCompletionResponse"
  /api/categories/{id}:
    put:
      tags:
      - category-controller
      operationId: updateCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    delete:
      tags:
      - category-controller
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/tasks:
    get:
      tags:
      - task-controller
      operationId: getAllTasks
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
    post:
      tags:
      - task-controller
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{id}/restore:
    post:
      tags:
      - task-controller
      operationId: restoreTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/sync/operations:
    get:
      tags:
      - sync-controller
      operationId: getOperations
      parameters:
      - name: since
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 0
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyncOperationResponse"
    post:
      tags:
      - sync-controller
      operationId: createOperation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncOperationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SyncOperationResponse"
  /api/habits:
    get:
      tags:
      - habit-controller
      operationId: getHabits
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitResponse"
    post:
      tags:
      - habit-controller
      operationId: createHabit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HabitResponse"
  /api/habit-completions:
    get:
      tags:
      - habit-completion-controller
      operationId: getCompletions
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitCompletionResponse"
    post:
      tags:
      - habit-completion-controller
      operationId: createCompletion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitCompletionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HabitCompletionResponse"
  /api/categories:
    get:
      tags:
      - category-controller
      operationId: getCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags:
      - category-controller
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
        required: true
      responses:
        "200":
          description: OK
  /api/auth/logout-all:
    post:
      tags:
      - auth-controller
      operationId: logoutAll
      responses:
        "200":
          description: OK
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/tasks/trash:
    get:
      tags:
      - task-controller
      operationId: getTrash
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
    delete:
      tags:
      - task-controller
      operationId: emptyTrash
      responses:
        "200":
          description: OK
  /api/tasks/{id}/permanent:
    delete:
      tags:
      - task-controller
      operationId: permanentlyDeleteTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
components:
  schemas:
    TaskRequest:
      required:
      - title
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        categoryId:
          type: string
        deadline:
          type: string
        completed:
          type: boolean
        wontDo:
          type: boolean
        priority:
          type: string
        imagePath:
          type: string
        updatedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        deviceId:
          type: string
    TaskResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        categoryId:
          type: string
        deadline:
          type: string
        completed:
          type: boolean
        wontDo:
          type: boolean
        priority:
          type: string
        imagePath:
          type: string
        updatedAt:
          type: integer
          format: int64
        deleted:
          type: boolean
        deletedAt:
          type: integer
          format: int64
        userId:
          type: string
        version:
          type: integer
          format: int64
        lastModifiedDeviceId:
          type: string
    HabitRequest:
      required:
      - title
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        groupName:
          type: string
        frequency:
          type: string
        totalDays:
          type: integer
          format: int32
        color:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int32
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        deviceId:
          type: string
    HabitResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        groupName:
          type: string
        frequency:
          type: string
        totalDays:
          type: integer
          format: int32
        color:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int32
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        deleted:
          type: boolean
        deletedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        lastModifiedDeviceId:
          type: string
    HabitCompletionRequest:
      required:
      - habitId
      - periodKey
      type: object
      properties:
        id:
          type: string
        habitId:
          type: string
        periodKey:
          type: string
        completedAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        deviceId:
          type: string
    HabitCompletionResponse:
      type: object
      properties:
        id:
          type: string
        habitId:
          type: string
        userId:
          type: string
        periodKey:
          type: string
        completedAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        deleted:
          type: boolean
        deletedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        lastModifiedDeviceId:
          type: string
    CategoryRequest:
      required:
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        pinned:
          type: boolean
        pinnedOrder:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int32
        hidden:
          type: boolean
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        deviceId:
          type: string
    CategoryResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        pinned:
          type: boolean
        pinnedOrder:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int32
        hidden:
          type: boolean
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        deleted:
          type: boolean
        deletedAt:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        lastModifiedDeviceId:
          type: string
    SyncOperationRequest:
      required:
      - deviceId
      - entityId
      - entityType
      - operation
      type: object
      properties:
        deviceId:
          type: string
        entityType:
          type: string
          enum:
          - task
          - category
          - habit
          - habit_completion
        entityId:
          type: string
        operation:
          type: string
          enum:
          - create
          - update
          - delete
          - restore
        baseVersion:
          type: integer
          format: int64
          minimum: 0
          description: Version observed by the client before creating the operation; use 0 for create or unknown.
        clientTimestamp:
          type: integer
          format: int64
          minimum: 0
        payloadJson:
          type: string
    SyncOperationResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        deviceId:
          type: string
        entityType:
          type: string
          enum:
          - task
          - category
          - habit
          - habit_completion
        entityId:
          type: string
        operation:
          type: string
          enum:
          - create
          - update
          - delete
          - restore
        baseVersion:
          type: integer
          format: int64
          minimum: 0
          description: Version observed by the client before creating the operation.
        clientTimestamp:
          type: integer
          format: int64
        serverTimestamp:
          type: integer
          format: int64
        payloadJson:
          type: string
    RegisterRequest:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 3
          type: string
        email:
          maxLength: 255
          minLength: 0
          type: string
        password:
          maxLength: 100
          minLength: 6
          type: string
        deviceId:
          type: string
    AuthResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: integer
          format: int64
        token:
          type: string
        tokenType:
          type: string
        expiresAt:
          type: integer
          format: int64
        refreshToken:
          type: string
        refreshExpiresAt:
          type: integer
          format: int64
        deviceId:
          type: string
        sessionId:
          type: string
    RefreshTokenRequest:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
        deviceId:
          type: string
    LogoutRequest:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
    LoginRequest:
      required:
      - identifier
      - password
      type: object
      properties:
        identifier:
          type: string
        password:
          maxLength: 100
          minLength: 6
          type: string
        deviceId:
          type: string