# backend/src/main/java/com/example/tisu/dto/LoginRequest.java
package com.example.tisu.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
        @NotBlank
        String identifier,

        @NotBlank
        @Size(min = 6, max = 100)
        String password,

        String deviceId
) {
}